Copy images and docs from main repo to wiki repo
Prior to this commit, many of the images and docs that the wiki pages were
referencing were hosted in the main OSCC repo instead of the wiki's
repo, which cluttered the main repo with unnecessary images. This commit
copies the images into the wiki repo and updates the references
accordingly.