Update firmware command and report descriptions
Prior to this commit, the commands and reports were described in terms
of bit locations which is not clear because one system may expect bits
to be in a certain location while another expects them somewhere else
(endianness). This commit changes the bits to be C data types with their
sizes.